+++ backend/packages/eks/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @cat-factory/eks
 
+## 0.1.252
+
+### Patch Changes
+
+- Updated dependencies [6076cf1]
+- Updated dependencies [2fdb08d]
+- Updated dependencies [11a2966]
+  - @cat-factory/kernel@0.256.0
+  - @cat-factory/contracts@0.258.0
+  - @cat-factory/integrations@0.138.2
+
 ## 0.1.251
 
 ### Patch Changes